// src/app/models/legacy/User.ts
export interface LegacyUser {
  DateofBirth: string;
  apiKey: string;
  connect_broker_status: string;
  country_code: number;
  created_at: Date;
  ddpi_enabled: boolean;
  edis: boolean;
  edis_updated_at: Date;
  email: string;
  firebase_id: string;
  image_url: string;
  is_authorized_for_sell: boolean;
  jwtToken: string;
  name: string;
  panNumber: string;
  phone_number: number;
  profile_completion: number;
  telegram_id: string;
  token_expire: Date;
  tpin_enabled: boolean;
  updated_at: Date;
  user_broker: string;
  user_onBoard_from: string;
  secretKey: string;
  clientCode: string;
  _id: string;
}